tests: Setup test config and provide structure for testing
